Added detection of npm5 fixing rid link, using npm install over npm link
This is needed since something has changed in npm5 that makes link fail,
probably a bug that will get addressed in the future.

Bug or not, this is the recommended way to work with monorepos in npm5
which basically is what we are doing here, so a good change anyhow.